error checking and finl touches

+++ controllers/userController.js
@@ -8,15 +8,27 @@ var bcrypt = require("bcrypt");
 
 // view
 exports.profile = asyncHandler(async (req, res, next) => {
-  const user = await User.findById(req.params.id).exec();
-  const blogResult = await Blog.find({user: user._id}).exec();
-  const commentsResult = await Comment.find({user: user._id}).exec();
+  try {
+    const user = await User.findById(req.params.id).exec();
+    const blogResult = await Blog.find({user: user._id}).exec();
+    const commentsResult = await Comment.find({user: user._id}).exec();
 
-  const commentBlogs = [];
-  for (const comment in commentsResult) {
-    commentBlogs.push(await Blog.findById(commentsResult[comment].blog).exec());
+    const commentBlogs = [];
+    for (const comment in commentsResult) {
+      commentBlogs.push(await Blog.findById(commentsResult[comment].blog).exec());
+    }
+    res.render("pages/profile", {
+      title: 'User profile',
+      user: req.user,
+      profile: user,
+      blogs: blogResult,
+      comments: commentsResult,
+      commentBlogs: commentBlogs,
+      errors: [req.query.errors]
+    });
+  } catch (e) {
+    res.redirect('/error/?code=500&errors=' + e);
   }
-  res.render("pages/profile", {title: 'User profile', user: req.user, profile: user, blogs: blogResult, comments: commentsResult, commentBlogs: commentBlogs, errors: [req.query.errors]});
 });
 
 // create
@@ -47,18 +59,19 @@ exports.registerP = [
 
         await user.save();
 
-        const { image } = req.files;
+        if (req.files != null && 'image' in req.files) {
+          const { image } = req.files;
 
-        if (image != undefined) {
           image.mv(process.cwd() + '/public' + user.image);
         }
 
         res.redirect(user.url);
       } catch (e) {
-        errors.push("can't create user")
+        console.log(e);
+        errors.push("Server error can't create user")
       }
     } else {
-      errors.push("invalid username or password")
+      errors.push("your password and username need to be at least 6 characters long")
     }
 
     res.render("pages/register", {title: 'make a new user', user: req.user, errors: errors});
@@ -92,12 +105,12 @@ exports.loginP = [
 
     const user = await User.findOne({username: req.body.username}).exec();
 
-
     if (!errors.isEmpty()) {
-      res.render("login", {errors: errors.array()});
+      res.render("pages/login", {title: 'Login', user: undefined, errors: errors.array()});
+      console.log(errors.array()[0].msg);
       return;
     } else if (user == null) {
-      res.render("login", {errors: ['Cant find user']});
+      res.render("pages/login", {title: 'Login', user: undefined, errors: ['Cant find user']});
       return;
     } else if (bcrypt.compareSync(req.body.password, user.password)) {
 
@@ -110,7 +123,7 @@ exports.loginP = [
 
       res.redirect('/');
     } else {
-      res.render("login", {errors: ['Wrong password']});
+      res.render("pages/login", {title: 'Login', user: undefined, errors: ['Wrong password']});
       return;
     }
   }),
@@ -130,8 +143,8 @@ exports.userDeleteG = asyncHandler(async (req, res, next) => {
     await User.deleteOne({_id: req.user._id}).exec();
 
     res.redirect('/');
-  } catch(e) {
-
+  } catch (e) {
+    res.redirect('/error/?code=500&errors=' + e);
   }
 
 });
@@ -148,39 +161,50 @@ exports.userUpdateP = [
     .trim()
     .isLength({ min: 6 })
     .escape(),
+  body("password_new")
+    .trim()
+    .isLength({ min: 6 })
+    .escape(),
   // do the stuff
   asyncHandler(async (req, res, next) => {
     const errors = validationResult(req).array();
-    var url;
+    let url = '';
 
     if (errors.length == 0) {
       if (req.user != undefined) {
         if (String(req.user._id) == req.params.id) {
-          try {
-            const user = new User({
-              username: req.body.username,
-              password: bcrypt.hashSync(req.body.password, 8),
-              _id: req.params.id,
-            });
+          const userold = await User.findOne({username: req.body.username}).exec();
 
-            const userup = await User.findByIdAndUpdate(req.params.id, user, {});
+          if (bcrypt.compareSync(req.body.password, userold.password)) {
+            try {
+              const user = new User({
+                username: req.body.username,
+                password: bcrypt.hashSync(req.body.password_new, 8),
+                _id: req.params.id,
+              });
+
+              const userup = await User.findByIdAndUpdate(req.params.id, user, {});
 
 
-            if (req.files != null && 'image' in req.files) {
-              const { image } = req.files;
+              if (req.files != null && 'image' in req.files) {
+                const { image } = req.files;
 
-              try {
-                const fs = require('fs');
-                fs.unlinkSync(process.cwd() + '/public' + userup.image);
-              } catch (e) {}
+                try {
+                  const fs = require('fs');
+                  fs.unlinkSync(process.cwd() + '/public' + userup.image);
+                } catch (e) {}
 
-              image.mv(process.cwd() + '/public' + userup.image);
+                image.mv(process.cwd() + '/public' + userup.image);
+              }
+
+
+              url = '/users/' + req.params.id;
+
+            } catch (e) {
+              url = '/users/' + req.params.id + "/?errors=cant update user";
             }
-
-
-            url = '/users/' + req.params.id;
-          } catch (e) {
-            url = '/users/' + req.params.id + "/?errors=cant update user";
+          } else {
+            url = '/users/' + req.params.id + "/?errors=password is wrong";
           }
         } else {
           url = '/users/' + req.params.id + "/?errors=you dont have permission to update this profile"
